Split duration card and forward starting
